Delete info about branch "Release"
This is because of my misunderstanding of "Releases", since I did not notice I can upload binaries directly. So I thought that I needed to create a separate branch to create a release, but I was wrong, so I deleted the branch "Release" and also deleted any information in README that pointed to this branch.